update: separate styled components from ChatBox
