--- src/components/Chat/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ChatPage'
