+++ src/components/ChatBox/index.ts
@@ -1,0 +1,2 @@
+export * from './ChatBox';
+export { default } from './ChatBox';
